Improve setup and assertions in extn functional tests
Ensure all exceptions are output with the full message and backtrace.

Parameterize harness to make it easier to copy and paste.

Relocate the `Kernel#require` functionals to the
`extn::core::kernel::require` module.
