+++ artichoke-backend/src/extn/core/thread/mod.rs
@@ -31,16 +31,20 @@ mod tests {
     use crate::test::prelude::*;
     use bstr::ByteSlice;
 
+    const SUBJECT: &str = "Thread";
+    const FUNCTIONAL_TEST: &[u8] = include_bytes!("thread_test.rb");
+
     #[test]
-    fn integration_test() {
+    fn functional() {
         let mut interp = crate::interpreter().unwrap();
-        let _ = interp.eval(&include_bytes!("thread_test.rb")[..]).unwrap();
+        let _ = interp.eval(FUNCTIONAL_TEST).unwrap();
         let result = interp.eval(b"spec");
         if let Err(exc) = result {
             let backtrace = exc.vm_backtrace(&mut interp);
             let backtrace = bstr::join("\n", backtrace.unwrap_or_default());
             panic!(
-                "Thread tests failed with message: '{:?}' and backtrace:\n{:?}",
+                "{} tests failed with message: {:?} and backtrace:\n{:?}",
+                SUBJECT,
                 exc.message().as_bstr(),
                 backtrace.as_bstr()
             );
