+++ artichoke-backend/src/extn/stdlib/json/mod.rs
@@ -37,15 +37,27 @@ pub struct Json;
 #[cfg(test)]
 mod tests {
     use crate::test::prelude::*;
+    use bstr::ByteSlice;
+
+    const SUBJECT: &str = "JSON";
+    const FUNCTIONAL_TEST: &[u8] = include_bytes!("json_test.rb");
 
     #[test]
     // TODO(GH-528): fix failing tests on Windows.
     #[cfg_attr(target_os = "windows", should_panic)]
-    fn integration_test() {
+    fn functional() {
         let mut interp = crate::interpreter().unwrap();
-        let _ = interp.eval(&include_bytes!("json_test.rb")[..]).unwrap();
+        let _ = interp.eval(FUNCTIONAL_TEST).unwrap();
         let result = interp.eval(b"spec");
-        let result = result.unwrap().try_into::<bool>(&interp).unwrap();
-        assert!(result);
+        if let Err(exc) = result {
+            let backtrace = exc.vm_backtrace(&mut interp);
+            let backtrace = bstr::join("\n", backtrace.unwrap_or_default());
+            panic!(
+                "{} tests failed with message: {:?} and backtrace:\n{:?}",
+                SUBJECT,
+                exc.message().as_bstr(),
+                backtrace.as_bstr()
+            );
+        }
     }
 }
