+++ artichoke-backend/src/extn/stdlib/strscan/mod.rs
@@ -13,13 +13,25 @@ pub struct StringScanner;
 #[cfg(test)]
 mod tests {
     use crate::test::prelude::*;
+    use bstr::ByteSlice;
+
+    const SUBJECT: &str = "StringScanner";
+    const FUNCTIONAL_TEST: &[u8] = include_bytes!("strscan_test.rb");
 
     #[test]
-    fn integration_test() {
-        let mut interp = crate::interpreter().expect("init");
-        interp.eval(&include_bytes!("strscan_test.rb")[..]).unwrap();
+    fn functional() {
+        let mut interp = crate::interpreter().unwrap();
+        let _ = interp.eval(FUNCTIONAL_TEST).unwrap();
         let result = interp.eval(b"spec");
-        let result = result.unwrap().try_into::<bool>(&interp).unwrap();
-        assert!(result);
+        if let Err(exc) = result {
+            let backtrace = exc.vm_backtrace(&mut interp);
+            let backtrace = bstr::join("\n", backtrace.unwrap_or_default());
+            panic!(
+                "{} tests failed with message: {:?} and backtrace:\n{:?}",
+                SUBJECT,
+                exc.message().as_bstr(),
+                backtrace.as_bstr()
+            );
+        }
     }
 }
