+++ artichoke-backend/src/extn/stdlib/uri/mod.rs
@@ -47,13 +47,25 @@ pub struct Uri;
 #[cfg(test)]
 mod tests {
     use crate::test::prelude::*;
+    use bstr::ByteSlice;
+
+    const SUBJECT: &str = "URI";
+    const FUNCTIONAL_TEST: &[u8] = include_bytes!("uri_test.rb");
 
     #[test]
-    fn integration_test() {
+    fn functional() {
         let mut interp = crate::interpreter().unwrap();
-        let _ = interp.eval(&include_bytes!("uri_test.rb")[..]).unwrap();
+        let _ = interp.eval(FUNCTIONAL_TEST).unwrap();
         let result = interp.eval(b"spec");
-        let result = result.unwrap().try_into::<bool>(&interp).unwrap();
-        assert!(result);
+        if let Err(exc) = result {
+            let backtrace = exc.vm_backtrace(&mut interp);
+            let backtrace = bstr::join("\n", backtrace.unwrap_or_default());
+            panic!(
+                "{} tests failed with message: {:?} and backtrace:\n{:?}",
+                SUBJECT,
+                exc.message().as_bstr(),
+                backtrace.as_bstr()
+            );
+        }
     }
 }
